move line chart model logic into own file
