+++ src/scripts/chart/line.js
@@ -4,14 +4,8 @@ var numberAxis = require('../axis/number.js');
 var textArea = require('../element/text-area.js');
 var lineKey = require('../element/line-key.js');
 var ftLogo = require('../element/logo.js');
-var lineThickness = require('../util/line-thickness.js');
 var interpolator = require('../util/line-interpolators.js');
-var ratios = require('../util/aspect-ratios.js');
-var seriesOptions = require('../util/series-options.js');
-
-function isDate(d) {
-	return d && d instanceof Date && !isNaN(+d);
-}
+var lineModel = require('./line.model.js');
 
 function getHeight(selection) {
 	return Math.ceil(selection.node().getBoundingClientRect().height);
@@ -21,172 +15,13 @@ function getWidth(selection) {
 	return Math.ceil(selection.node().getBoundingClientRect().width);
 }
 
-function translate(margin) {
-	return function(position) {
-		var left = position.left || 0;
-		var top = position.top || 0;
-		return 'translate(' + (margin + left) + ',' + top + ')';
-	};
-}
-
 function lineChart(p) {
     'use strict';
-
-	var lineClasses = ['series1', 'series2', 'series3', 'series4', 'series5', 'series6', 'series7', 'accent'];
-
-	//these classes can be used in addition to those above
-	var complementaryLineCLasses = ['forecast'];
-
-	function buildModel(opts) {
-		var m = {
-			//layout stuff
-			height: undefined,
-			width: 300,
-			chartHeight: undefined,
-			chartWidth: undefined,
-			blockPadding: 8,
-			simpleDate: false,
-			simpleValue: false,
-			logoSize: 28,
-			//data stuff
-			falseorigin: false, //TODO, find out if there's a standard 'pipeline' temr for this
-			error: function(err) { console.log('ERROR: ', err); },
-			lineClasses: {},
-			niceValue: true,
-			hideSource: false,
-			numberAxisOrient: 'left',
-			margin: 2,
-			lineThickness: undefined,
-			x: {
-				series: '&'
-			},
-			y: {
-				series: []
-			},
-			labelLookup: null,
-			sourcePrefix: 'Source: '
-		};
-
-		for (var key in opts) {
-			m[key] = opts[key];
-		}
-
-		m.contentWidth = m.width - (m.margin * 2);
-
-		m.translate = translate(0);
-
-		// only set the chart width if hasn't been defined
-		// explicitly in the config
-		if (!m.chartWidth) {
-			// minus gutter for logo
-			var rightGutter = m.contentWidth < 260 ? 16 : 26;
-			m.chartWidth = m.contentWidth - rightGutter;
-		}
-
-		if (!m.chartHeight) {
-			// The chart size should have a nice aspect ratio
-			var isNarrow = m.chartWidth < 220;
-			var isWide = m.chartWidth > 400;
-			var ratio = isNarrow ? 1.1 : (isWide ? ratios.commonRatios.widescreen : ratios.commonRatios.standard);
-			m.chartHeight = ratios.heightFromWidth(m.chartWidth, ratio);
-		}
-
-		m.lineStrokeWidth = lineThickness(m.lineThickness);
-
-		m.x.series = seriesOptions.normalise(m.x.series);
-		m.y.series = seriesOptions.normaliseY(m.y.series)
-							.filter(function (d) {
-								return !!d.key && d.key !== m.x.series.key;
-							})
-							.map(function (d, i) {
-								d.index = i;
-								d.className = lineClasses[i];
-								return d;
-							});
-
-		if (typeof m.key !== 'boolean') {
-			m.key = m.y.series.length > 1;
-		} else if (m.key && !m.y.series.length) {
-			m.key = false;
-		}
-
-		m.data = !Array.isArray(m.data) ? [] : m.data.map(function (d, i) {
-
-			var s = d[m.x.series.key];
-			var error = {
-				node: null,
-				message: '',
-				row: i,
-				column: m.x.series.key,
-				value: s
-			};
-
-			if (!d) {
-				error.message = 'Empty row';
-			} else if (!s) {
-				error.message = 'X axis value is empty or null';
-			} else if (!isDate(s)) {
-				error.message = 'Value is not a valid date';
-			}
-
-			if (error.message) {
-				m.error(error);
-				d[m.x.series.key] = null;
-			}
-
-			return d;
-
-		});
-
-		//make sure all the lines are numerical values, calculate extents...
-		var extents = [];
-		m.y.series.forEach(function (l, i) {
-			var key = l.key;
-			m.data = m.data.map(function (d, j) {
-
-				var value = d[key];
-				var isValidNumber = value === null || typeof value === 'number';
-				if (!isValidNumber) {
-					m.error({
-						node: null,
-						message: 'Value is not a number',
-						value: value,
-						row: j,
-						column: key
-					});
-				}
-				return d;
-			});
-			var ext = d3.extent(m.data, function(d){
-				return d[key];
-			});
-			extents = extents.concat (ext);
-		});
-
-		//work out the time domain
-		if (!m.timeDomain) {
-			m.timeDomain = d3.extent(m.data, function (d) {
-				return d[m.x.series.key];
-			});
-		}
-
-		//work out the value domain
-		if (!m.valueDomain) {
-			m.valueDomain = d3.extent(extents);
-			// unless a false origin has been specified
-			if (!m.falseorigin && m.valueDomain[0] > 0) {
-				m.valueDomain[0] = 0;
-			}
-		}
-
-		return m;
-	}
 
 	function chart(g){
 
 		//the model is built froma  copy of the data
-		var model = buildModel(Object.create(g.data()[0]));
-
+		var model = lineModel.build(Object.create(g.data()[0]));
 		var svg = g.append('svg')
 				.attr({
 					'class': 'graphic line-chart',
@@ -194,8 +29,8 @@ function lineChart(p) {
 					height: model.height,
 					width: model.width
 				});
-		var titleMetaData = svg.append('title').text(model.title);
-		var subtitleMetaData = svg.append('desc').text(model.subtitle);
+		svg.append('title').text(model.title);
+		svg.append('desc').text(model.subtitle);
 
 		var defaultLineHeight = 1.2;
 		// TODO: don't hard-code the fontsize, get from CSS somehow.
